requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
diffusers==0.27.2
torch==2.2.0
torchvision==0.17.0
transformers==4.39.0
accelerate==0.28.0
numpy==1.26.4
pillow==10.3.0
scipy==1.13.0
pydantic==2.7.0
